+++ canvas-component-tools/CHANGELOG.md
@@ -8,6 +8,7 @@ Canvas Component Tools uses [semantic versioning](https://semver.org/).
 * Updated template titles to not use 404 title to avoid confusion
 * Updated command to output version of the tool instead of Canvas
 * Updated dependencies
+* Removed `presets` array from global schema template
 * Fixed path to `getCanvasConfig()` helper function
 
 ## 1.0.0 - 2022-11-03