+++ canvas-component-tools/CHANGELOG.md
@@ -1,6 +1,16 @@
-# 📅 Canvas Component Tools Changelog
+# 📚 Canvas Component Tools Changelog
 
 Canvas Component Tools uses [semantic versioning](https://semver.org/).
+
+## 1.1.0 - 2022-31-01
+
+* Added limited interactivity to template options
+* Added documentation links to footer as choices are made
+* Updated template titles to not use 404 title to avoid confusion
+* Updated command to output version of the tool instead of Canvas
+* Updated dependencies
+* Removed `presets` array from global schema template
+* Fixed path to `getCanvasConfig()` helper function
 
 ## 1.0.0 - 2022-11-03